+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## Version 0.0.1
+- Refactor Configs object to be singleton instead of the managers (#4)
+- Refactor Firebase manager's batabase query actions into a DAO (#4)
 - Added crud services and api endpoints for questions (#3)
 - Set up Firebase Firestore and its emulator (#2)
 - Added database schemas (#1)